server/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
--no-binary psycopg2

argon2-cffi==18.3.0
cffi==1.11.5              # via argon2-cffi
click==6.7                # via pip-tools
django-debug-toolbar==1.9.1
django-environ==0.4.5
django-extensions==2.0.7
django-redis==4.9.0
django==2.0.7
djangorestframework==3.8.2
first==2.0.1              # via pip-tools
gunicorn==19.9.0
pip-tools==2.0.2
psycopg2==2.7.5
pycparser==2.18           # via cffi
pytz==2018.5              # via django
redis==2.10.6             # via django-redis
six==1.11.0               # via argon2-cffi, django-extensions, pip-tools
sqlparse==0.2.4           # via django-debug-toolbar
